requirements.txt
torch==2.10.0
lightning==2.6.1
wandb==0.25.1
hydra-core==1.3.2
pytest